+++ tests/test_audio.py
@@ -1,6 +1,5 @@
 from utils.audiokit import sound
 import numpy as np
-import unittest
 
 wavObj = sound.Wave(
     channelsPerFrame=1,
@@ -12,16 +11,16 @@ wavObj = sound.Wave(
     wavName="sine.wav"
 )
 
-class TestSoundMethods(unittest.TestCase):
+def test_generate_samples_and_writes_and_plays_samples_as_audio() -> None:
+    
+    sampleArray = wavObj.generateSamples()
+    wavObj.writeWav(wavData=sampleArray)
+    wavObj.playAudio(audioData=sampleArray)
+#end def test
 
-    def test_sine_at_t_0(self):
-        sineResult = wavObj.sineWave(time=0)
-        self.assertEqual(sineResult, 0)
-        self.assertFalse(sineResult != 0)
-
-    def test_generate_samples(self):
-        sampleArray = wavObj.generateSamples()
-        self.assertEqual(sampleArray.dtype, np.dtype('int16'))
-
-if __name__ == '__main__':
-    unittest.main()
+# Run Testing
+# In root directory, run python3 -m tests.test_audio
+if __name__ == "__main__":
+    test_generate_samples_and_writes_and_plays_samples_as_audio()
+    print("Audio Testing Passed!")
+# end if
